% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{checkOrCreateCluster}
\alias{checkOrCreateCluster}
\title{Create a cluster inside pnd-specific functions (internal use only)}
\usage{
checkOrCreateCluster(cl = NULL, cores = 1)
}
\arguments{
\item{cl}{\code{NULL} = return a PSOCK cluster with exported objects on Windows
or the \code{'lapply' / 'mclapply'} string on everything else}

\item{cores}{If 1, return \code{'lapply'}; if more, return a cluster on Windows
or \code{'mclapply'} on everything else.}
}
\value{
\code{'lapply'}, \code{'mclapply'}, or a cluster object.
}
\description{
Create a cluster inside pnd-specific functions (internal use only)
}
\examples{
cl <- checkOrCreateCluster(cores = 1)  # "lapply"
cl <- tryCatch(checkOrCreateCluster("mclapply 2"), error = function(e) return(NULL))
print(cl)
}
