% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.stepsize}
\alias{plot.stepsize}
\title{Step-size selection visualisation}
\usage{
\method{plot}{stepsize}(x, ...)
}
\arguments{
\item{x}{List returned by \code{step...} functions.}

\item{...}{Additional graphical parameters passed to \code{plot()}.}
}
\value{
Nothing (invisible null).
}
\description{
Plots the estimated truncation error and total errors, highlighting various
ranges obtained during step-size selection for numerical differentiation.
Works for all implemented methods.
}
\examples{

sCR <- step.CR(sin, 1)
sK <- step.K(sin, 1)
plot(sCR)
plot(sK)
}
