% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{setupParallelEnv}
\alias{setupParallelEnv}
\title{Create an environment for PSOCK parallelisation}
\usage{
setupParallelEnv(FUN, cl, ...)
}
\arguments{
\item{FUN}{A function to be called within an environment.}

\item{cl}{A cluster to which the environment is exported.}

\item{...}{Arguments passed to FUN.}
}
\value{
A list with a function, environment, and dot arguments.
}
\description{
Create an environment for PSOCK parallelisation
}
