% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{runParallel}
\alias{runParallel}
\title{Run a function in parallel over a list (internal use only)}
\usage{
runParallel(FUN, x, cores = 1, preschedule = FALSE, cl = NULL)
}
\arguments{
\item{FUN}{A function of only one argument. If there are more arguments, use
the \code{FUN2 <- do.call(FUN, c(list(x), ...))} annd call it.}

\item{x}{A list to parallelise the evaluation of \code{FUN} over.}

\item{cores}{Integer specifying the number of CPU cores used for parallel computation.
Recommended to be set to the number of physical cores on the machine minus one.}

\item{preschedule}{Logical: if \code{TRUE}, disables pre-scheduling for \code{mclapply()}
or enables load balancing with \code{parLapplyLB()}. Recommended for functions that
take less than 0.1 s per evaluation.}

\item{cl}{An optional user-supplied \code{cluster} object (created by \code{makeCluster}
or similar functions). If not \code{NULL}, the code uses \code{parLapply()}
(if \code{preschedule} is \code{TRUE}) or \code{parLapplyLB()} on that cluster on Windows
and \code{mclapply} (fork cluster) on everything else.}
}
\description{
Run a function in parallel over a list (internal use only)
}
